Implement manual operator assignment
