+++ back/docker-service-discovery/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "docker_discovery"
-version = "0.1.3"
+version = "0.1.4"
 description = "This package provides a dummy implementation for the ServiceDiscoveryPort based on reading a docker-compose.yml file."
 authors = [
     {name = "Fernando González Rivero",email = "fernando.gonzalezr@upb.edu.co"}
